+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## next release
 ### Features
 ### Improvements
+
+* Bump `requests` to `>=2.31.0` circumvent `CVE-2023-32681`
+
 ### Bugfix