+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Bugfix
 
+* Fix error handling of FieldManager if no mapped source field exists in the event.
 
 ## v6.3.0
 ### Features